+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [1.11.0]
-* Added `BottomResizablePane` widget
+* 🚨 Breaking Changes 🚨
+* `ResizablePane.startWidth` has been changed to `ResizablePane.startSize`
+* `ResizablePane.minWidth` has been changed to `ResizablePane.minSize`
+* `ResizablePane.maxWidth` has been changed to `ResizablePane.maxSize`
+* `ResizablePane` can now be vertically resized
 
 ## [1.10.0]
 * Added implementation of `MacosDisclosureButton`